+++ R/runners.R
@@ -74,7 +74,6 @@ make_method <- function(
 #' @param alpha Optional. Significance level to pass to the engine.
 #' @param score Optional. A scoring function for score-based methods.
 #' @param ... Additional arguments passed to the engine-specific runner.
-#' @param directed_as_undirected_knowledge Logical. Used internally for \pkg{pcalg}.
 #'
 #' @return An object representing a configured runner for the chosen engine.
 #'   The type depends on the engine.
@@ -87,8 +86,7 @@ make_runner <- function(
   test = NULL,
   alpha = NULL,
   score = NULL,
-  ...,
-  directed_as_undirected_knowledge = FALSE
+  ...
 ) {
   if (exists(engine, envir = engine_registry_env, inherits = FALSE)) {
     custom_engine <- engine_registry_env[[engine]]
@@ -103,8 +101,7 @@ make_runner <- function(
       test = test,
       alpha = alpha,
       score = score,
-      ...,
-      directed_as_undirected_knowledge = directed_as_undirected_knowledge
+      ...
     ))
   }
 
@@ -144,8 +141,7 @@ make_runner <- function(
       test = test,
       alpha = alpha,
       score = score,
-      ...,
-      directed_as_undirected_knowledge = directed_as_undirected_knowledge
+      ...
     ),
     bnlearn = .make_runner_bnlearn(
       alg = alg,
@@ -170,8 +166,7 @@ make_runner <- function(
   test = NULL,
   alpha = NULL,
   score = NULL,
-  ...,
-  directed_as_undirected_knowledge = FALSE
+  ...
 ) {
   engine_map <- list(
     causalDisco = list(class = CausalDiscoSearch, pkgs = "causalDisco"),
@@ -225,10 +220,7 @@ make_runner <- function(
 
   list(
     set_knowledge = function(knowledge) {
-      search$set_knowledge(
-        knowledge,
-        directed_as_undirected = directed_as_undirected_knowledge
-      )
+      search$set_knowledge(knowledge)
     },
     run = function(data) {
       search$run_search(data)