+++ inst/roxygen-examples/order_restrict_amat_cpdag-example.R
@@ -14,8 +14,7 @@ ss_cor <- make_suff_stat(tpc_example, type = "cor_test")
 wrapped_test <- dir_test(cor_test, vnames, kn)
 pc_cons <- .pcalg_constraints_from_knowledge(
   kn,
-  labels = vnames,
-  directed_as_undirected = TRUE
+  labels = vnames
 )
 
 set.seed(1405)
@@ -25,6 +24,5 @@ skel <- pcalg::skeleton(
   alpha = 0.05,
   labels = vnames,
   method = "stable.fast",
-  fixedGaps = pc_cons$fixedGaps,
-  fixedEdges = pc_cons$fixedEdges
+  fixedGaps = pc_cons$fixed_gaps
 )